+++ .zed/settings.json
@@ -1,0 +1,12 @@
+{
+  "lsp": {
+    "rust-analyzer": {
+      "initialization_options": {
+        "cargo": {
+          "target": "thumbv8m.main-none-eabihf",
+          "allTargets": false
+        }
+      }
+    }
+  }
+}
