+++ crates/drivers/lsm6dso-ner/Cargo.toml
