+++ crates/drivers/vl6180x-ner/src/error.rs
