+++ crates/drivers/vl6180x-ner/src/mode/powered_off.rs
