+++ projects/README.md
@@ -1,0 +1,5 @@
+This is the location of all of the board-specific and high level operating code.
+
+Each folder is a project corresponding to ONE binary.
+
+Feel free to depend on code inside of crates/drivers folder.
