+++ projects/cerberus/Cargo.toml
@@ -1,7 +1,10 @@
+cargo-features = ["per-package-target"]
+
 [package]
 edition = "2021"
 name = "cerberus"
 version = "0.1.0"
+forced-target = "thumbv7em-none-eabi"
 
 [dependencies]
 cortex-m.workspace = true
@@ -10,7 +13,7 @@ defmt.workspace = true
 defmt-rtt.workspace = true
 embassy-embedded-hal.workspace = true
 embassy-executor.workspace = true
-embassy-stm32.workspace = true
+embassy-stm32 = { workspace = true, features = ["stm32f405rg"] }
 embassy-sync.workspace = true
 embassy-time.workspace = true
 embassy-futures.workspace = true
@@ -18,4 +21,4 @@ heapless.workspace = true
 panic-probe.workspace = true
 static_cell.workspace = true
 bitfield.workspace = true
-pca9539-ner = { version = "0.1.0", path = "../crates/pca9539-ner" }
+pca9539-ner = { version = "0.1.0", path = "../../crates/drivers/pca9539-ner" }
