+++ projects/lightning/.cargo/config.toml
@@ -1,0 +1,2 @@
+[build]
+target = "thumbv8m.main-none-eabihf"
