+++ projects/lightning/Cargo.toml
@@ -1,7 +1,10 @@
+cargo-features = ["per-package-target"]
+
 [package]
-edition = "2021"
-name = "wheel"
+edition = "2024"
+name = "lightning"
 version = "0.1.0"
+forced-target = "thumbv7em-none-eabi"
 
 [dependencies]
 cortex-m.workspace = true
@@ -10,7 +13,7 @@ defmt.workspace = true
 defmt-rtt.workspace = true
 embassy-embedded-hal.workspace = true
 embassy-executor.workspace = true
-embassy-stm32.workspace = true
+embassy-stm32 = { workspace = true, features = ["stm32h563zi"] }
 embassy-sync.workspace = true
 embassy-time.workspace = true
 embassy-futures.workspace = true
