+++ projects/wheel/Cargo.toml
@@ -1,0 +1,22 @@
+cargo-features = ["per-package-target"]
+
+[package]
+edition = "2021"
+name = "wheel"
+version = "0.1.0"
+forced-target = "thumbv8m.main-none-eabihf"
+
+[dependencies]
+cortex-m.workspace = true
+cortex-m-rt.workspace = true
+defmt.workspace = true
+defmt-rtt.workspace = true
+embassy-embedded-hal.workspace = true
+embassy-executor.workspace = true
+embassy-stm32 = { workspace = true, features = ["stm32f405rg"] }
+embassy-sync.workspace = true
+embassy-time.workspace = true
+embassy-futures.workspace = true
+heapless.workspace = true
+panic-probe.workspace = true
+#static_cell.workspace = true
