+++ rust-toolchain.toml
@@ -1,5 +1,5 @@
 # The channel key should match the stable rust-toolchain.toml in the embassy repo
 [toolchain]
-channel = "1.92"
-components = [ "rust-src", "rustfmt", "llvm-tools", "clippy" ]
-targets = [ "thumbv7em-none-eabi" ]
+channel = "nightly-2025-12-11"
+components = [ "rust-src", "rustfmt", "llvm-tools", "clippy", "miri" ]
+targets = [ "thumbv7em-none-eabi", "thumbv8m.main-none-eabihf" ]
